:truck: Move: 프로젝트 페이지 폴더 구조 변경
Move: 프로젝트 페이지 폴더 구조 변경
