+++ app/project/[projectId]/_components/MyComment/WriteComment/WriteButton.tsx
@@ -6,7 +6,7 @@ import Button from "@/app/_components/Button/Button";
 import { commentQueryKeys } from "@/app/_queryFactory/commentQuery";
 import { useToast } from "@/app/_context/ToastContext";
 import { projectQueryKeys } from "@/app/_queryFactory/projectQuery";
-import { useEnterCommentContext } from "../../_context/EnterCommentProvider";
+import { useEnterCommentContext } from "../../../_context/EnterCommentProvider";
 
 interface Props {
   projectId: number;
