+++ app/project/[projectId]/_components/Project/ImageSlide/ButtonSlide.tsx
